refactor: Extract ActivityFeed step cards and contact details into dedicated components.
